+++ LibraryProject/src/Vector2/Vector2.h
@@ -4,7 +4,7 @@
 class Vector2
 {
 public:
-	Vector2(float x, float y);
+	Vector2(double x, double y);
 	Vector2();
 
 	~Vector2() = default;
